requirements.txt
Django>=4.2
django-allauth
requests
openfga-sdk>=0.9.5
python-dotenv
PyJWT
cryptography
sentence_transformers
langchain
faiss-cpu
